Update server to split into routers
