+++ client/src/App.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from "react";
+import React from "react";
 import "./App.css";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import LearningPlan from "./screens/LearningPlan";
@@ -6,7 +6,7 @@ import LoginScreen from "./screens/login/login";
 import Register from "./screens/register/Register";
 import DefaultScreen from "./screens/Default";
 import ProtectedRoute from "./components/ProtectedRoutes/ProtectedRoute";
-import { AuthProvider, useAuth } from "./context/AuthContext";
+import { AuthProvider } from "./context/AuthContext";
 
 function App() {
   return (
