+++ server/controller/loginProcessor.js
@@ -4,7 +4,7 @@ require('dotenv').config({
 });
 const bcrypt = require("bcrypt");
 const DatabaseParser = require("../parser/databaseParser");
-const STATUS_CODES = require("../statusCodes");
+const STATUS_CODES = require("../utils/statusCodes");
 const StatusCodes = require("./StatusCodes");
 
 class LoginAPI {
