+++ server/routes/profileRoutes.js
@@ -1,0 +1,20 @@
+const express = require("express");
+const router = express.Router();
+const authenticateToken = require("../utils/authenticateToken");
+const initializeErrorMessages = require("../utils/errorMessages");
+const ProfileAPI = require("../controller/profileProcessor");
+
+const ERROR_MESSAGES = initializeErrorMessages();
+const profileAPI = new ProfileAPI();
+
+router.get('/get/:id', authenticateToken, async(req, res) => {
+    const profileQuery = await profileAPI.getProfile(req.body.email);
+    if(typeof profileQuery !== "object") {
+        console.error("There was a problem retrieving profile.");
+        res.status(profileQuery).send(ERROR_MESSAGES.get(profileQuery));
+        return;
+    }
+    res.status(STATUS_CODES.OK).json(profileQuery);
+});
+
+module.exports = router;
