+++ server/tests/controller/loginProcessor.test.js
@@ -1,7 +1,7 @@
 const bcrypt = require("bcrypt");
 const LoginAPI = require("../../controller/loginProcessor");
 const DatabaseParser = require("../../parser/databaseParser");
-const STATUS_CODES = require("../../statusCodes");
+const STATUS_CODES = require("../../utils/statusCodes");
 
 jest.mock("../../parser/DatabaseParser", () => {
     const testParser = {
