+++ server/tests/controller/profileProcessor.test.js
@@ -1,6 +1,6 @@
 const ProfileAPI = require("../../controller/profileProcessor")
 const DatabaseParser = require("../../parser/databaseParser");
-const STATUS_CODES = require("../../statusCodes");
+const STATUS_CODES = require("../../utils/statusCodes");
 
 jest.mock("../../parser/DatabaseParser", () => {
     const testParser = {
